Fixes part #3
* Refactoring for different view binding strategies
* Added notification balloone
